starting new db strucutre , this will allow us to manage 1 table for all site / server combos

+++ database/migrations/2016_11_04_174711_add_key_and_cert_ssl_tables.php
@@ -31,14 +31,18 @@ class AddKeyAndCertSslTables extends Migration
      */
     public function down()
     {
-        Schema::table('server_ssl_certificates', function (Blueprint $table) {
-            $table->dropColumn('key');
-            $table->dropColumn('cert');
-        });
+        if(Schema::hasTable('server_ssl_certificates')) {
+            Schema::table('server_ssl_certificates', function (Blueprint $table) {
+                $table->dropColumn('key');
+                $table->dropColumn('cert');
+            });
+        }
 
-        Schema::table('site_ssl_certificates', function (Blueprint $table) {
-            $table->dropColumn('key');
-            $table->dropColumn('cert');
-        });
+        if(Schema::hasTable('site_ssl_certificates')) {
+            Schema::table('site_ssl_certificates', function (Blueprint $table) {
+                $table->dropColumn('key');
+                $table->dropColumn('cert');
+            });
+        }
     }
 }
